feat(tests): replace generic tests with comprehensive landing page tests
